Chapter5

+++ src/items/items.service.ts
@@ -1,8 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateItemDto } from './dto/create-item.dto';
 import { ItemStatus } from './item-status.enum';
 import { Item } from './../entities/item.entity';
 import { ItemRepository } from './item.repository';
+import { User } from 'src/entities/user.entity';
 @Injectable()
 export class ItemsService {
   constructor(private readonly itemRepository: ItemRepository) {}
@@ -12,8 +13,8 @@ export class ItemsService {
     return await this.itemRepository.find();
   }
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
-    return await this.itemRepository.createItem(createItemDto);
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
+    return await this.itemRepository.createItem(createItemDto, user);
   }
 
   async findById(id: string): Promise<Item> {
@@ -23,14 +24,22 @@ export class ItemsService {
     }
     return found;
   }
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id);
+    if (item.userId === user.id) {
+      throw new BadRequestException(
+        '自身で登録した商品のステータスは変更できません',
+      );
+    }
     item.status = ItemStatus.SOLD_OUT;
     return await this.itemRepository.save(item);
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
     const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません');
+    }
     this.itemRepository.delete(item);
   }
 }
